+++ tests/test_sign_tx_cmd.py
@@ -117,7 +117,7 @@ def test_sign_tx_with_tx_source_enabled(backend, scenario_navigator, device, nav
     assert response == expected_signature
 
 
-def test_sign_tx_with_nested_authorization_disabled(
+def test_sign_tx_with_authorization_details_disabled(
     backend, scenario_navigator, device, navigator
 ):
     keypair = Keypair.from_mnemonic_phrase(MNEMONIC, index=0)
@@ -128,7 +128,7 @@ def test_sign_tx_with_nested_authorization_disabled(
     configure_device_settings(
         navigator,
         device,
-        SettingsId.DISABLE_NESTED_AUTHORIZATION | SettingsId.ENABLE_BLIND_SIGNING,
+        SettingsId.DISABLE_AUTHORIZATION_DETAILS | SettingsId.ENABLE_BLIND_SIGNING,
     )
 
     with client.sign_tx(path=path, transaction=signature_base):